refactored execution into specific subcommand executors
Before, the specific methods were called "prepare_commandline_...", as the idea was that they'd just assemble
the command, which would then be executed, after being returned, at a top level. While initially ok, with more
and more of the subcommands, this loop was broken (terminated by exec in prepare_commandline). This
refactoring cleans that confusion up and more properly names the executors "exec_cmd_...".