allow_abbrev parser arg removed to allow flag combinations, fixes #3
It seems that the new allow_abbrev parser arg not only controls long
option abbreviations (which we actually don't want), but also hinders
flag combinations (such as `-i -t` --> `-it`). As the latter is very
common, i rate  allowing abbreviations (while docker itself doesn't) as
minor (if not a feature).